backend/requirements.txt
Django>=4.2,<5.0
python-dotenv
psycopg2-binary
dj-database-url
djangorestframework
supabase
django-cors-headers
pypdf==6.7.4
anthropic>=0.39.0
langchain-anthropic>=0.3.0
pydantic>=2.0
pdf2image>=1.16.0
django-filter>=24.3
Pillow>=10.0.0
requests>=2.31.0
